refactor: isolate namespace AgentFS surfaces
