+++ crates/agent-engine/src/runtime/agent_loop/namespace_environment/tests.rs
@@ -3,6 +3,7 @@ use std::sync::{
     atomic::{AtomicUsize, Ordering as AtomicOrdering},
 };
 
+use alan_agent_protocol::{ContentPart, InputMode, Op};
 use alan_agentfs::{AgentConformanceChecker, AgentFs, AgentRootFs};
 use alan_ap::{
     ErrorCode, Fid, FileKind, FileServer, InProcessTransport, OpenMode, Qid, Request, Stat,
