+++ crates/agent-engine/src/runtime/agent_loop/namespace_environment/tests/tape.rs
@@ -1,3 +1,4 @@
+use super::super::agent_files::tape_record_bytes;
 use super::*;
 
 #[tokio::test]
